backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==4.0.1
boto3==1.42.58
botocore==1.42.58
click==8.3.1
colorama==0.4.6
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.133.1
h11==0.16.0
httptools==0.7.1
idna==3.11
jmespath==1.1.0
motor==3.7.1
passlib==1.7.4
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
PyJWT==2.11.0
pymongo==4.16.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
PyYAML==6.0.3
s3transfer==0.16.0
six==1.17.0
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
watchfiles==1.1.1
websockets==16.0
google-api-python-client==2.162.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
